Added a ThreadedSubsystemTemplate. Made GlobalCoordinatePosition and LinearSlide inherit from that. Made the LED class. It will inherit from that.
