+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/src/Utills/ThreadedSubsystemTemplate.java
@@ -3,6 +3,8 @@ package org.firstinspires.ftc.teamcode.src.Utills;
 public abstract class ThreadedSubsystemTemplate implements Runnable {
     protected volatile boolean isRunning = true;
     protected long sleepTime = 50;
+    protected Executable<Boolean> opModeIsActive;
+    protected Executable<Boolean> isStopRequested;
 
     protected abstract void threadMain();
 
@@ -11,14 +13,19 @@ public abstract class ThreadedSubsystemTemplate implements Runnable {
     }
 
     public void run() {
-        while (isRunning) {
-            this.threadMain();
-            try {
+        try {
+            while (isRunning && !isStopRequested.call()) {
+                this.threadMain();
                 Thread.sleep(sleepTime);
-            } catch (InterruptedException e) {
-                return;
             }
+        } catch (NullPointerException | InterruptedException e) {
+            e.printStackTrace();
         }
+
+    }
+
+    public boolean isRunning() {
+        return this.isRunning;
     }
